+++ src/components/SignInPage/SignInPage.styled.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-export const Form = styled.form`
+export const SignInForm = styled.form`
   height: fit-content;
 
   display: flex;
@@ -13,4 +13,8 @@ export const Form = styled.form`
     gap: 10px;
   }
 `;
-export const SignInPage = styled.div``;
+export const SignInPage = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+`;
